SuggestionService added

+++ c_weather/src/main/java/hr/evolaris/air/foi/evolaris_smarttourism/c_weather/Main.java
@@ -11,7 +11,7 @@ public class Main
 
     @SerializedName("pressure")
     @Expose
-    public Integer pressure;
+    public Double pressure;
 
     @SerializedName("humidity")
     @Expose
