+++ c_weather/src/main/java/hr/evolaris/air/foi/evolaris_smarttourism/c_weather/Wind.java
@@ -10,5 +10,5 @@ public class Wind
     public Double speed;
     @SerializedName("deg")
     @Expose
-    public Integer deg;
+    public Double deg;
 }
